requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
fake-useragent==1.5.1
Pillow==10.3.0
Flask-Mail==0.10.0
gunicorn==22.0.0
gunicorn==22.0.0
